+++ tomcat/src/test/java/org/apache/coyote/http/request/RequestTest.java
@@ -1,11 +1,7 @@
-package org.apache.coyote.http.request.exception;
+package org.apache.coyote.http.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.coyote.http.request.HttpRequestBody;
-import org.apache.coyote.http.request.HttpRequestHeaders;
-import org.apache.coyote.http.request.QueryParameters;
-import org.apache.coyote.http.request.Url;
 import org.apache.coyote.http.util.HttpMethod;
 import org.apache.coyote.http.util.HttpVersion;
 import org.junit.jupiter.api.DisplayNameGeneration;
